(PC-17812) favorites: rename pages/utils to pages/helpers
